Extract communication actions
